fix leave team

+++ app/controllers/members_controller.rb
@@ -1,12 +1,16 @@
 class MembersController < ApplicationController
   def destroy
-    member = Member.find(params[:id])
-    this_member = current_user.member
-    member.destroy
-    if this_member == member
-      redirect_to root_url
+    if !current_user.member.captain?
+      member = Member.find(params[:id])
+      this_member = current_user.member
+      member.destroy
+      if this_member == member
+        redirect_to root_url
+      else
+        render json: { status: :ok, member: { id: member.id } }
+      end
     else
-      render json: { status: :ok, member: { id: member.id } }
+      redirect_to my_team_url
     end
   end
